+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Feat: Allow setting proguard via Options and/or external resources (#1728)
 * Feat: Add breadcrumbs for the Apollo integration (#1726)
 * Fix: Don't set lastEventId for transactions (#1727)